refactor(chain): split EVMChainAdapter god-class via mixin pattern
Phase 2 of the evm-adapter.ts structural split. Decompose the 4372-line
EVMChainAdapter class into a shared base + per-domain mixin holders, mirroring
the DKGAgent split. Pure mechanical move — no behaviour or public-API change.

- evm-adapter-base.ts          all instance state, constructor, tx/Hub
                               plumbing, and helpers shared across groups
                               (private members widened to protected)
- evm-adapter-identity.ts      identity / profile / operational-wallet
- evm-adapter-publish.ts       KC publish / update / mint pipeline
- evm-adapter-context-graph.ts context-graph create / publish / read / policy
- evm-adapter-conviction.ts    publishing-conviction account / agent
- evm-adapter-ack-sign.ts      ACK / signature / identity verification
- evm-adapter-random-sampling.ts  challenge / proof
- evm-adapter-storage-reads.ts low-level on-chain reads
- evm-adapter-events.ts        listenForEvents
- evm-adapter-constants.ts     module-scope timing / key-purpose constants
- evm-adapter-apply-mixins.ts  applyMixins helper

evm-adapter.ts now only assembles the class (extends base + applyMixins) and
re-exports the public API. Holder methods carry no `this:` annotation, so the
composed method signatures stay byte-identical to the pre-split class and
external callers (incl. EVMChainAdapter | NoChainAdapter unions) are unaffected;
the four cross-group helpers were promoted to base to keep holders self-contained.

evm-adapter.ts: 4372 -> 52 lines.
