+++ packages/chain/src/evm-adapter-storage-reads.ts
@@ -1,0 +1,72 @@
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Low-level on-chain storage read methods.
+ *
+ * Mixin holder extracted from evm-adapter.ts. `extends EVMChainAdapterBase`
+ * for shared state (providers, signers, caches) reached via `this`. Bodies
+ * are a 1:1 move — no behaviour change. Mixed into the concrete EVMChainAdapter
+ * via applyMixins(); see evm-adapter.ts for the assembly.
+ */
+
+import { EVMChainAdapterBase } from './evm-adapter-base.js';
+import { Contract, ethers } from 'ethers';
+
+export class StorageReadMethods extends EVMChainAdapterBase {
+  // =====================================================================
+  // KC views (V10 DKGKnowledgeAssets + ContextGraphStorage)
+  // =====================================================================
+
+  requireKCStorage(): Contract {
+    const kas = this.contracts.knowledgeAssetStorage;
+    if (!kas) {
+      throw new Error(
+        'DKGKnowledgeAssets not deployed in this Hub. ' +
+        'V10 KC views require a Hub with DKGKnowledgeAssets registered.',
+      );
+    }
+    return kas;
+  }
+
+  async getLatestMerkleRoot(kaId: bigint): Promise<Uint8Array> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const rootHex: string = await kas.getLatestMerkleRoot(kaId);
+    return ethers.getBytes(rootHex);
+  }
+
+  async getMerkleLeafCount(kaId: bigint): Promise<number> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const count: bigint = BigInt(await kas.getMerkleLeafCount(kaId));
+    return Number(count);
+  }
+
+  async getLatestCiphertextChunksRoot(kaId: bigint): Promise<Uint8Array> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const rootHex: string = await kas.getLatestCiphertextChunksRoot(kaId);
+    return ethers.getBytes(rootHex);
+  }
+
+  async getCiphertextChunkCount(kaId: bigint): Promise<number> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const count: bigint = BigInt(await kas.getCiphertextChunkCount(kaId));
+    return Number(count);
+  }
+
+  async getLatestMerkleRootPublisher(kaId: bigint): Promise<string> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const publisher: string = await kas.getLatestMerkleRootPublisher(kaId);
+    return publisher;
+  }
+
+  async getLatestMerkleRootAuthor(kaId: bigint): Promise<string> {
+    await this.init();
+    const kas = this.requireKCStorage();
+    const author: string = await kas.getLatestMerkleRootAuthor(kaId);
+    return author;
+  }
+}
